Instant update for the UI when publishing a post
