+++ src/components/home/Home.jsx
@@ -10,6 +10,7 @@ import SearchPage from "../search/SearchPage.jsx";
 import IconButton from '@mui/material/Button';
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import Divider from '@mui/material/Divider';
+import { getPublicPosts } from '../../APIRequests'
 
 // https://stackoverflow.com/a/64517088
 export const AuthorIdContext = React.createContext({
@@ -31,6 +32,35 @@ export default function Home() {
   const [authorId, setAuthorId] = useState(localStorage.getItem("authorId") ? localStorage.getItem("authorId") : userId);
   const [showSearch, setShowSearch] = useState(localStorage.getItem("showSearch") ? localStorage.getItem("showSearch") : "false");
   const [query, setQuery] = useState(localStorage.getItem("query") ? localStorage.getItem("query") : "");
+
+  // Code for PublishTab and Myposts
+  const size = 5;
+  const [numPages, setNumPages] = React.useState(0);
+  const [page, setPage] = React.useState(1);
+
+  const handlePostsChange = (event, value) => {
+    setPage(value);
+    updateMyPosts(value, size);
+  };
+
+  const [inbox, setInbox] = React.useState([]);
+
+  const updateMyPosts = (page, size) => {
+    // State change will cause component re-render
+    async function fetchPublicPosts() {
+      const output = await getPublicPosts(authorId, page, size);
+      if (output.posts.length == 0) { return; }
+      setInbox(output.posts);
+      setNumPages(Math.ceil(output.count/size));
+    }
+    fetchPublicPosts();
+  }
+
+  React.useEffect(() => {
+    updateMyPosts(page, size);
+    // disable this warning because updateMyPosts has to be used outside of the useEffect as well
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   // https://stackoverflow.com/a/53455443
   const handleBack = () => {
@@ -62,7 +92,12 @@ export default function Home() {
                 <Grid item xs={3}>
                   <Grid container rowSpacing={3} sx={{ width: '90%', marginRight: 3, marginLeft: 3, marginTop: 3}}>
                     <Grid item xs={12} align="center" justify="top">
-                      <PublishButton/>
+                      <PublishButton 
+                        updateMyPosts={updateMyPosts} 
+                        page={page} size={size} 
+                        handlePostsChange={handlePostsChange} 
+                        setInbox={setInbox}
+                        />
                     </Grid>
                     <Grid item xs={12} align="center" justify="center">
                       <Divider orientation="horizontal" flexItem sx={{ mr: "-1px" }} />
@@ -74,7 +109,12 @@ export default function Home() {
                 </Grid>
                 <Divider orientation="vertical" flexItem sx={{ mr: "-1px", minHeight: 700}} />
                 <Grid item xs={5.5}>
-                  <HomeTab authorId={userId}/>
+                  <HomeTab 
+                    authorId={userId} 
+                    inbox={inbox} 
+                    numPages={numPages} 
+                    handlePostsChange={handlePostsChange}
+                    />
                 </Grid>
                 <Divider orientation="vertical" flexItem sx={{ mr: "-1px", minHeight: 700 }} />
                 <Grid item xs={3.5}>
