+++ src/components/homeTab/HomeTab.jsx
@@ -41,7 +41,7 @@ export default function TabsWrappedLabel(props) {
           value === "one" && <Inbox authorId={props.authorId}/>
         }
         {
-          value === "two" && <MyPosts authorId={props.authorId}/>
+          value === "two" && <MyPosts authorId={props.authorId} inbox={props.inbox} numPages={props.numPages} handlePostsChange={props.handlePostsChange}/>
         }
         {
           value === "three" && <GithubEvents github={author.github}/>
